refactor(runtime): split value/error.rs into cohesive modules (§7-8)
`value/error.rs` was 1179 lines, over the 500-line module limit. Split it into
three files along its natural seams, each under 500 lines, with no behavior
change (the set of fn signatures is identical before/after):

- `error.rs` (471) — the `RuntimeErrorCode` enum + `Control` enum + the
  `RuntimeError` struct definition, `expected_type_object`, the `is_*` control
  predicates, `from_exception_value`/`with_location`, all control-flow signal
  constructors (`return_signal`, `last_signal`, `take_signal`, …), and the
  `RuntimeErrorCode` unit tests.
- `error_construct.rs` (237) — `RuntimeError` constructors for numeric, range,
  and assignment errors (divide-by-zero, `out_of_range`, `assignment_ro`,
  `typed`/`typed_msg`, `str_numeric`, …).
- `error_typed.rs` (484) — `RuntimeError` constructors for
  declaration/syntax/typecheck errors (`undeclared*`, `redeclaration*`,
  `method_not_found`, `syntax_*`, `typecheck_*`, `x_adverb`, …) plus JSON
  exception rendering (`to_json_exception` + `json_string`/`json_value`).

All `RuntimeError` constructors are already `pub(crate)`, so no visibility change
is needed; `error_typed` imports the `pub(crate)` `expected_type_object` free
fn. Each file carries only the imports it uses (`error_typed` drops the unused
`Symbol`; `error.rs` drops the now-unused `HashMap`). Ninth slice of the §7-8
file-splitting sweep.
